add /files endpoints
